General cleanup of error messages and redirect.
The page will now redirect more quickly, however, I am still unable to find a way to force it to redirect without loading anything at all. The error messages will now appear in a "modal", which is far easier on the eyes and won't interrupt page flows when an irrelevant error occurrs.

Closes #4