Added support for IE and tested.
IE6 works but renders terribly. Anyone on Windows XP may upgrade to IE8, which works. IE7, which was only used on Vista, has a minor rendering issue; this is from Bootstrap and is probably not fixable by us. IE8 and IE10 work as expected. IE9 is being tested, and I will re-open this issue if there is a problem.

Closes #17