improve printing pagination for US Letter pages
the recent graph6 changes increased the height of the graph causing
it to spill onto the third page (when printing the page as a pdf).

this change refactors the svg to create groups <g> for both graphs
5 & 6 and for the two of them as page2 and then applies a transform
to avoid changing all the individual elements.

Note this change worsens printing on A4.